// apps/nest-api/src/modules/auth/users/user.interface.ts
import { EStatus } from '../../../interfaces/status.interface';

export enum EGender {
    MALE = 'MALE',
    FEMALE = 'FEMALE',
}

export enum ERole {
    ADMIN = 'ADMIN',
    USER = 'USER',
}

export interface IUser {
    id: string;
    cpf: string;
    role: ERole;
    salt?: string;
    name: string;
    email: string;
    gender?: EGender;
    status: EStatus;
    picture?: string;
    password?: string;
    createdAt: Date;
    updatedAt?: Date;
    deletedAt?: Date;
    dateOfBirth?: Date;
    recoverToken?: string;
    confirmationToken?: string;
}